refactor: rename to NuThoughtsSettingsTab
